+++ Drivers/BSP/Components/hx8347i/hx8347i.h
@@ -49,8 +49,8 @@ typedef int32_t (*HX8347I_Init_Func)    (void);
 typedef int32_t (*HX8347I_DeInit_Func)  (void);
 typedef int32_t (*HX8347I_GetTick_Func) (void);
 typedef int32_t (*HX8347I_Delay_Func)   (uint32_t);
-typedef int32_t (*HX8347I_WriteReg_Func)(uint16_t, uint16_t, uint8_t*, uint16_t);
-typedef int32_t (*HX8347I_ReadReg_Func) (uint16_t, uint16_t, uint8_t*, uint16_t);
+typedef int32_t (*HX8347I_WriteReg_Func)(uint16_t, uint16_t, uint8_t*, uint32_t);
+typedef int32_t (*HX8347I_ReadReg_Func) (uint16_t, uint16_t, uint8_t*, uint32_t);
 
 typedef struct
 {
