+++ Drivers/BSP/Components/hx8347i/hx8347i_reg.h
@@ -175,8 +175,8 @@ extern "C" {
 /** @defgroup HX8347I_REG_Exported_Types HX8347I_REG Exported Types
   * @{
   */ 
-typedef int32_t (*HX8347I_Write_Func)(void *, uint16_t, uint8_t*, uint16_t);
-typedef int32_t (*HX8347I_Read_Func) (void *, uint16_t, uint8_t*, uint16_t);
+typedef int32_t (*HX8347I_Write_Func)(void *, uint16_t, uint8_t*, uint32_t);
+typedef int32_t (*HX8347I_Read_Func) (void *, uint16_t, uint8_t*, uint32_t);
 
 typedef struct
 {
@@ -191,8 +191,8 @@ typedef struct
 /** @defgroup HX8347I_REG_Exported_Functions HX8347I_REG Exported Functions
   * @{
   */ 
-int32_t hx8347i_write_reg(HX8347I_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint16_t length);
-int32_t hx8347i_read_reg(HX8347I_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint16_t length);
+int32_t hx8347i_write_reg(HX8347I_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint32_t length);
+int32_t hx8347i_read_reg(HX8347I_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint32_t length);
 /**
   * @}
   */
