+++ Drivers/BSP/Components/st7789h2/st7789h2_reg.c
@@ -42,7 +42,7 @@
 * Input         : Driver context, register Address, length of buffer
 * Output        : Status.
 *******************************************************************************/
-int32_t st7789h2_read_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint16_t length)
+int32_t st7789h2_read_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint32_t length)
 {
   return ctx->ReadReg(ctx->handle, reg, pdata, length);
 }
@@ -54,7 +54,7 @@ int32_t st7789h2_read_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uin
 *                 length of buffer.
 * Output        : Status.
 *******************************************************************************/
-int32_t st7789h2_write_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint16_t length)
+int32_t st7789h2_write_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, uint32_t length)
 {
   return ctx->WriteReg(ctx->handle, reg, pdata, length);
 }
@@ -65,7 +65,7 @@ int32_t st7789h2_write_reg(ST7789H2_ctx_t *ctx, uint16_t reg, uint8_t *pdata, ui
 * Input         : Driver context, data to be written, length of buffer.
 * Output        : Status.
 *******************************************************************************/
-int32_t st7789h2_send_data(ST7789H2_ctx_t *ctx, uint8_t *pdata, uint16_t length)
+int32_t st7789h2_send_data(ST7789H2_ctx_t *ctx, uint8_t *pdata, uint32_t length)
 {
   return ctx->SendData(ctx->handle, pdata, length);
 }
