+++ Drivers/BSP/STM32L562E-DK/stm32l562e_discovery_lcd.c
@@ -120,9 +120,9 @@ static void    ST7789H2_PowerDown(void);
 
 static int32_t LCD_FMC_Init(void);
 static int32_t LCD_FMC_DeInit(void);
-static int32_t LCD_FMC_WriteReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint16_t Length);
-static int32_t LCD_FMC_ReadReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint16_t Length);
-static int32_t LCD_FMC_Send(uint8_t *pData, uint16_t Length);
+static int32_t LCD_FMC_WriteReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint32_t Length);
+static int32_t LCD_FMC_ReadReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint32_t Length);
+static int32_t LCD_FMC_Send(uint8_t *pData, uint32_t Length);
 static int32_t LCD_FMC_GetTick(void);
 static void    FMC_MspInit(SRAM_HandleTypeDef *hSram);
 static void    FMC_MspDeInit(SRAM_HandleTypeDef *hSram);
@@ -919,10 +919,10 @@ static int32_t LCD_FMC_DeInit(void)
   * @param  Length  Number of data.
   * @retval BSP status.
   */
-static int32_t LCD_FMC_WriteReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint16_t Length)
+static int32_t LCD_FMC_WriteReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint32_t Length)
 {
   int32_t  ret = BSP_ERROR_NONE;
-  uint16_t i = 0;
+  uint32_t i = 0;
 
   if ((DevAddr != LCD_FMC_ADDRESS) || (pData == NULL) || (Length == 0U))
   {
@@ -953,10 +953,10 @@ static int32_t LCD_FMC_WriteReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData
   * @param  Length  Number of data.
   * @retval BSP status.
   */
-static int32_t LCD_FMC_ReadReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint16_t Length)
+static int32_t LCD_FMC_ReadReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData, uint32_t Length)
 {
   int32_t  ret = BSP_ERROR_NONE;
-  uint16_t i = 0;
+  uint32_t i = 0;
   uint16_t tmp;
 
   if ((DevAddr != LCD_FMC_ADDRESS) || (pData == NULL) || (Length == 0U))
@@ -987,10 +987,10 @@ static int32_t LCD_FMC_ReadReg16(uint16_t DevAddr, uint16_t Reg, uint8_t *pData,
   * @param  Length  Number of data.
   * @retval BSP status.
   */
-static int32_t LCD_FMC_Send(uint8_t *pData, uint16_t Length)
+static int32_t LCD_FMC_Send(uint8_t *pData, uint32_t Length)
 {
   int32_t  ret = BSP_ERROR_NONE;
-  uint16_t i = 0;
+  uint32_t i = 0;
 
   if ((pData == NULL) || (Length == 0U))
   {
